add more use x result types, tidy up the folder structure
