+++ packages/react-client/src/hooks/internal/useFishjamContext.ts
@@ -3,7 +3,7 @@ import { createContext, type MutableRefObject, useContext } from "react";
 
 import type { DeviceManager } from "../../devices/DeviceManager";
 import type { TrackManager } from "../../types/internal";
-import type { BandwidthLimits, PeerStatus, ScreenshareApi } from "../../types/public";
+import type { BandwidthLimits, PeerStatus, UseScreenshareResult } from "../../types/public";
 import type { FishjamClientState } from "./useFishjamClientState";
 
 export type FishjamContextType = {
@@ -11,7 +11,7 @@ export type FishjamContextType = {
   videoDeviceManagerRef: MutableRefObject<DeviceManager>;
   audioDeviceManagerRef: MutableRefObject<DeviceManager>;
   hasDevicesBeenInitializedRef: MutableRefObject<boolean>;
-  screenShareManager: ScreenshareApi;
+  screenShareManager: UseScreenshareResult;
   peerStatus: PeerStatus;
   videoTrackManager: TrackManager;
   audioTrackManager: TrackManager;
