+++ packages/react-client/src/hooks/useConnection.ts
@@ -1,6 +1,7 @@
-import type { GenericMetadata } from "@fishjam-cloud/ts-client";
+import type { GenericMetadata, ReconnectionStatus } from "@fishjam-cloud/ts-client";
 import { useCallback } from "react";
 
+import type { PeerStatus } from "../types/public";
 import { useFishjamContext } from "./internal/useFishjamContext";
 import { useReconnection } from "./internal/useReconnection";
 
@@ -19,6 +20,29 @@ export interface JoinRoomConfig<PeerMetadata extends GenericMetadata = GenericMe
   peerMetadata?: PeerMetadata;
 }
 
+export interface UseConnectionResult {
+  /**
+   * Join room and start streaming camera and microphone
+   *
+   * See {@link JoinRoomConfig} for parameter list
+   */
+  joinRoom: <PeerMetadata extends GenericMetadata = GenericMetadata>(
+    config: JoinRoomConfig<PeerMetadata>,
+  ) => Promise<void>;
+  /**
+   * Leave room and stop streaming
+   */
+  leaveRoom: () => void;
+  /**
+   * Current peer connection status
+   */
+  peerStatus: PeerStatus;
+  /**
+   * Current reconnection status
+   */
+  reconnectionStatus: ReconnectionStatus;
+}
+
 /**
  * Hook allows to to join or leave a room and check the current connection status.
  * @category Connection
@@ -30,12 +54,8 @@ export function useConnection() {
 
   const reconnectionStatus = useReconnection();
 
-  const joinRoom = useCallback(
-    <PeerMetadata extends GenericMetadata = GenericMetadata>({
-      url,
-      peerToken,
-      peerMetadata,
-    }: JoinRoomConfig<PeerMetadata>) => client.connect({ url, token: peerToken, peerMetadata: peerMetadata ?? {} }),
+  const joinRoom: UseConnectionResult["joinRoom"] = useCallback(
+    ({ url, peerToken, peerMetadata }) => client.connect({ url, token: peerToken, peerMetadata: peerMetadata ?? {} }),
     [client],
   );
 
@@ -46,17 +66,8 @@ export function useConnection() {
   const peerStatus = context.peerStatus;
 
   return {
-    /**
-     * Join room and start streaming camera and microphone
-     *
-     * See {@link JoinRoomConfig} for parameter list
-     */
     joinRoom,
-    /**
-     * Leave room and stop streaming
-     */
     leaveRoom,
-
     peerStatus,
     reconnectionStatus,
   };
