+++ packages/react-client/src/index.ts
@@ -3,16 +3,15 @@ export { useCamera, type UseCameraResult } from "./hooks/devices/useCamera";
 export { useInitializeDevices, type UseInitializeDevicesResult } from "./hooks/devices/useInitializeDevices";
 export { useMicrophone, type UseMicrophoneResult } from "./hooks/devices/useMicrophone";
 export type { JoinRoomConfig } from "./hooks/useConnection";
-export { useConnection } from "./hooks/useConnection";
+export { useConnection, type UseConnectionResult } from "./hooks/useConnection";
 export { type PeerWithTracks, usePeers, type UsePeersResult } from "./hooks/usePeers";
-export { useScreenShare } from "./hooks/useScreenShare";
+export { useScreenShare, type UseScreenshareResult } from "./hooks/useScreenShare";
 export { useUpdatePeerMetadata } from "./hooks/useUpdatePeerMetadata";
 export { useVAD } from "./hooks/useVAD";
 export type {
   DeviceItem,
   PeerStatus,
   PersistLastDeviceHandlers,
-  ScreenshareApi,
   StreamConfig,
   Track,
   TrackMiddleware,
